feat: 다가오는 일정 세부 페이지의 모든 일정 버튼 제거
