+++ frontend/src/features/shared-schedule/ui/UpcomingSchedules/ControlButton.tsx
@@ -1,7 +1,34 @@
+import { Flex } from '@/components/Flex';
 import { ChevronLeft, ChevronRight } from '@/components/Icon';
 import { vars } from '@/theme/index.css';
 
 import { controlButtonStyle } from './index.css';
+
+const ControlButtons = ({ 
+  translateCarousel, 
+  canTranslateLeft,
+  canTranslateRight,
+}: {
+  translateCarousel: (direction: 'left' | 'right') => void;
+  canTranslateLeft: boolean;
+  canTranslateRight: boolean;
+}) => (
+  <Flex
+    align='center'
+    gap={400}
+    justify='flex-end'
+    width='full'
+  >
+    <LeftControlButton
+      isAvailable={canTranslateLeft}
+      onClick={() => translateCarousel('left')}
+    />
+    <RightControlButton
+      isAvailable={canTranslateRight}
+      onClick={() => translateCarousel('right')}
+    />
+  </Flex>
+);
 
 interface ControlButtonProps {
   isAvailable: boolean;
@@ -33,3 +60,5 @@ export const RightControlButton = ({ isAvailable, onClick }: ControlButtonProps)
     />
   </button>
 );
+
+export default ControlButtons;
