+++ frontend/src/pages/UpcomingSchedulePage/index.css.ts
@@ -5,7 +5,6 @@ import { vars } from '@/theme/index.css';
 
 export const containerStyle = style({
   width: '100%',
-  marginTop: 80,
   paddingBottom: 218,
 });
